First Aid: Initial bringup
After deleting a file that is necessary for the project to function the application no longer closes while trying to read it. It just does not include it in the list. An option is available in the Settings to repair these broken projects.

TODO: Allow a colour and a favicon to be chosen while repairing.