Project: Show warning while working with files
Pretty simple I guess. Shows warning that says that you do not need to enter a file ending. If the file is an image or font it says you will need to. Added an option to disable this warning in Settings. Also fixed a few strings and fixed a bug with the ActionBarDrawerToggle for projects with relatively dark colours.