The perfect design?
Perhaps not perfect. But it is a lot better than before.  Removed a lot of extraneous code and improved lots of design aspects. The extra things that need to be done before release is:

* Implement pushing and pulling to remotes
* Remove unused Strings from strings.xml
* Use ExpandedListView for file viewer instead of NavigationView due to limited API

After this then we may think about release. Also I need a way of automating updating JavaDoc.