Remove color attribute for projects
So here's my reasoning. Colour is only used for one thing (theme-color) which can be set manually. It could be used for layout elements but it won't always fit with the scheme. Also having an extra property and a color picker library takes space. I honestly think it is better this way.